Switch to fmParallel
Work around libplacebo’s thread-unsafety with a mutex instead of fmUnordered.
I don’t know jackshit about concurrent programming tbh,
so I have no idea if this is even remotely sane or correct,
but it seems to work lmao.
Provides nearly a 2x speed up for .Tonemap and .Shader,
and a solid speed-up of up to 1.2x for .Resample and .Deband.